requirements.txt
flask
python-dotenv
openai
google-genai
supabase
python-telegram-bot
httpx
gunicorn
